aula 20/05

+++ primeiro-site/resources/views/clients/index.blade.php
@@ -8,7 +8,13 @@
             <h2>Usuários</h2>
         </div>
         <div class="pull-right">
+            
+            <!--se tiver a permissão mostra caso contrário não mostar-->
+            @can('cliente-create')
+
             <a class="btn btn-success" href="{{ route('clients.create') }}"> + Novo Cliente</a>
+
+            @endcan
         </div>
     </div>
 </div>
@@ -45,15 +51,22 @@
     <td>{{ $cliente->endereco }}</td>
     <td>
         <a class="btn btn-info" href="{{ route('clients.show',$cliente->id) }}">Mostrar</a>
+
+        @can('cliente-edit')
+
         <a class="btn btn-primary" href="{{ route('clients.edit',$cliente->id) }}">Editar</a>
 
+        @endcan
+
+        @can('cliente-delete')
+        
          {!! Form::open(['method' => 'DELETE','route' => ['clients.destroy', $cliente->id],'style'=>'display:inline']) !!}
 
              {!! Form::submit('Apagar', ['class' => 'btn btn-danger']) !!}
 
          {!! Form::close() !!}
 
-
+        @endcan
 
     </td>
   </tr>
